started generalizing free applicative to work with
any functor. Currently Existential quantification must be supported. An
idea for this would be to ceate a type that wraps a high order type (for
now I'll work with Kind2 as this may give insite for the more general
KindN case). We would also need to provide a constructor for this that
allows creating inhabitions of this wrapper as well as a few combinators
for applying functions to the wrapped value (although we must be careful
in that wrapped existential types cannot be unwrapped).

Existential <F,A> = <R>(cont : <E>(x : Kind2<F,A,E) => R) => R
mkExt : <F,A,E>(x : HKT<F,A,E>) : Existential<F, A>

                                     E = E' (maybe?)
                                        ┌────────────────┐
apExt : <F,A,B,E,E'>(f : (x : HKT<F, A, E>) => HKT<F, B, E'>) : Existential<F, B>